Restore depfile toleration of multiple output paths on distinct lines
Prior to introduction of depfile parser handling of multiple rules,
ninja silently accepted a depfile of the form:

    out: in1 in2 in3
    other: otherIn1 otherIn2 otherIn3

and incorrectly treated `other` and `otherIn*` as additional inputs to
`out`.  Now we prefer to reject this just as we already do for a depfile
specifying multiple outputs on one line.  However, this can break
existing cases where such a depfile was silently tolerated.

Add a `-w depfilemulti={err,warn}` option to control this behavior,
and make it just a warning by default.